requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18

# LangChain
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0

# Vector store
chromadb>=0.6.0

# File processing
PyMuPDF>=1.25.0
python-docx>=1.1.0
Pillow>=11.0.0

# Config
pyyaml>=6.0
pydantic>=2.0

# Utilities
httpx>=0.28.0
